fix: fix sonar issues
