+++ frontend/src/pages/Objects/components/MyReservations.tsx
@@ -3,41 +3,7 @@ import { Calendar, Clock, Eye, MapPin, User } from "lucide-react";
 import { Button } from "../../../components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "../../../components/ui/card";
 import { UserObjectReservation } from "../../../services/objects.ts";
-
-const REASON_PREVIEW_CHARS = 120;
-
-function normalizeReasonText(value: string): string {
-  const normalizedLines = value
-    .replace(/\r\n/g, "\n")
-    .split("\n")
-    .map((line) => line.trimEnd());
-
-  const compactLines: string[] = [];
-  let previousWasBlank = false;
-
-  for (const line of normalizedLines) {
-    const isBlank = line.trim().length === 0;
-    if (isBlank) {
-      if (!previousWasBlank) {
-        compactLines.push("");
-      }
-      previousWasBlank = true;
-      continue;
-    }
-
-    compactLines.push(line);
-    previousWasBlank = false;
-  }
-
-  return compactLines.join("\n").trim();
-}
-
-function buildReasonPreview(value: string, maxChars: number): string {
-  if (value.length <= maxChars) {
-    return value;
-  }
-  return `${value.slice(0, maxChars).trimEnd()}...`;
-}
+import { normalizeReasonText, buildReasonPreview, REASON_PREVIEW_CHARS, formatDateTime } from "../../../utils/rentalFormattingUtils";
 
 interface MyReservationsProps {
   reservations: UserObjectReservation[];
@@ -52,17 +18,9 @@ interface MyReservationsProps {
   onRetry: () => void;
 }
 
-function formatDateTime(dateString: string) {
-  const date = new Date(dateString);
-  return new Intl.DateTimeFormat("es-ES", {
-    dateStyle: "medium",
-    timeStyle: "short",
-  }).format(date);
-}
-
 function canCancelReservation(rental: any): boolean {
   return (
-    ["ACTIVE", "IN_PROGRESS"].includes(rental.status) &&
+    rental.status === "ACTIVE" &&
     new Date(rental.end_date) > new Date()
   );
 }
